% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoMbo.R
\name{autoMbo}
\alias{autoMbo}
\title{Benchmark and plot \code{mlrMBO::mbo()} optimization runs to investigate hyperparameters}
\usage{
autoMbo(
  data,
  target,
  minimize = FALSE,
  funcEvals = 45,
  minFuncEvals = 10,
  itersMboTune = 10,
  repls = 10,
  showInfo = TRUE,
  ncpus = NA,
  seed = 1
)
}
\description{
This functions benchmarks the \code{mlrMBO::mbo()} function on different configurations and
then plots them wrt the hyperparameters.
}
