% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMboHyperparams.R
\name{plotMboHyperparams}
\alias{plotMboHyperparams}
\title{plot of \code{mlrMBO::mbo()} optimization runs to investigate hyperparameters}
\usage{
plotMboHyperparams(
  task,
  funcEvals,
  psParamPlot,
  resolution,
  repls,
  showInfo = TRUE,
  ncpus = NA,
  seed = 1
)
}
\value{
A plot to investigate hyperparameters.
}
\description{
This functions benchmarks the \code{mlrMBO::mbo()} function on different configurations and
then plots them wrt the hyperparameters.
}
\examples{
\dontrun{
}

}
\references{
Bernd Bischl, Jakob Richter, Jakob Bossek, Daniel Horn, Janek Thomas and Michel Lang; mlrMBO: A Modular Framework for Model-Based Optimization of Expensive Black-Box Functions, Preprint: \code{\link{https://arxiv.org/abs/1703.03373}} (2017).
}
