% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task}
\alias{task}
\title{Generate task}
\usage{
task(simulation = "regr.randomForest", data, target, psOpt, minimize = TRUE)
}
\arguments{
\item{simulation}{[\code{character}]\cr
The black box function e.g. model for the \code{mlrMBO}
Default is `regr.randomForest`.}

\item{data}{[\code{data.frame}]\cr
Data of problem. \cr}

\item{target}{[\code{character}]\cr
The variable, which one wants to minimize (maximize) \code{mlrMBO}.}

\item{psOpt}{[\code{ParamHelpers::ParamSet()}]\cr
Collection of parameters and their constraints for problem optimization}

\item{minimize}{[\code{logical(1)}]\cr
Should the target be minimized? \cr
Default is `TRUE`.}
}
\value{
List with relevant informations of a specific problem.
}
\description{
This functions enables users to create a task, which mainly describes the optimization problem.
Task is required for many EBO functions, such as EBO::optimizertuneRace(), EBO::generateConfigdata()
or EBO::generateProblemList().
}
\examples{
\dontrun{
set.seed(1)

library(mlrMBO)
library(ParamHelpers)
library(mlr)

set.seed(1)

# define problem / task
data <- data.frame(a = runif(50,10,5555), b = runif(50,-30000,-500),
                   c = runif(50,0,1000))
data$ratio <- rowSums(data[,1:3]^2)
data$ratio <- data$ratio/max(data$ratio)
colnames(data) <- c("power", "time", "pressure","ratio")

psOpt = makeParamSet(

 makeIntegerParam("power", lower = 10, upper = 5555),

 makeIntegerParam("time", lower = -30000, upper = -500),

 makeIntegerParam("pressure", lower = 0, upper = 1000)
)

task = task(
 simulation = "regr.randomForest",
 data = data,
 target = "ratio",
 psOpt = psOpt,
 minimize = FALSE
)
}
}
