% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testAddIters.R
\name{testAddIters}
\alias{testAddIters}
\title{run statistical tests for \code{mlrMBO::mbo()} iterations}
\usage{
testAddIters(configResults, baseIters = 20, addIters = 10, minimize = TRUE)
}
\arguments{
\item{baseIters}{[\code{integer(1)}]\cr
Define the number of baseline iterations.\cr
Default is 10.}

\item{addIters}{[\code{integer(1)}]\cr
Define the number of additional iterations.\cr
Default is 10.}

\item{minimize}{[\code{logical(1)}]\cr
Should the target be minimized? \cr
Default is `TRUE`.}
}
\value{
Test results and summaries for effect of iterations.
}
\description{
This function benchmarks the \code{mlrMBO::mbo()} function on different iterations for different
configuration and run statistical tests for iterations.
}
\examples{
\dontrun{

set.seed(1)

library(mlrMBO)
library(ParamHelpers)
library(mlr)

#define infillCrit
ctrl = mlrMBO::makeMBOControl()
ctrl = mlrMBO::setMBOControlInfill(ctrl, crit = mlrMBO::makeMBOInfillCritEI())

# define MBO configuration
paramsMBO = data.table::data.table(
 design = list("maximinLHS","randomLHS", "random"),
 amountDesign = list(12),
 control = list(ctrl),
 surrogate = list(mlr::makeLearner("regr.km", predict.type = "se"))
)

namesBoxplot = c("maximinLHS",
                "randomLHS",
                "random")

# define runs of each algorithm
repls = 10

# define function evaluations
funcEvals = 32


data <- data.frame(a = runif(50,10,5555), b = runif(50,-30000,-500),
                   c = runif(50,0,1000))
data$ratio <- rowSums(data[,1:3]^2)
data$ratio <- data$ratio/max(data$ratio)
colnames(data) <- c("power", "time", "pressure","ratio")


psOpt = ParamHelpers::makeParamSet(
  ParamHelpers::makeIntegerParam("power", lower = 10, upper = 5555),
  ParamHelpers::makeIntegerParam("time", lower = -30000, upper = -500),
  ParamHelpers::makeNumericParam("pressure", lower = 0, upper = 1000),
)

task = task(
 simulation = "regr.randomForest",
 data = data,
 target = "ratio",
 psOpt = psOpt,
 minimize = FALSE
)

# generate configData
configResults = generateConfigdata(task, funcEvals = funcEvals, paramsMBO,
                                       namesBoxplot = namesBoxplot, repls = repls)

# analyze effect of additional iterations
testAddIters(configResults, baseIters = 10, addIters = 10, minimize = FALSE)
}
}
\references{
[\code{mlrMBO::mbo()}]

Bernd Bischl, Jakob Richter, Jakob Bossek, Daniel Horn, Janek Thomas and Michel Lang; mlrMBO: A Modular Framework for Model-Based Optimization of Expensive Black-Box Functions, Preprint: \code{\link{https://arxiv.org/abs/1703.03373}} (2017).
}
\seealso{
\code{\link{optimize::plotBenchmark()}} \code{\link{optimize::plotMboContourPlot()}}
}
