% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBenchmark.R
\name{plotBenchmark}
\alias{plotBenchmark}
\title{Benchmark and plot all optimizers which are implemented in EBO.}
\usage{
plotBenchmark(
  instance,
  psOpt,
  funcEvals = 65,
  paramsMBO = data.table::data.table(NULL),
  paramsCMAESR = data.table::data.table(NULL),
  paramsES = data.table::data.table(NULL),
  paramsDE = data.table::data.table(NULL),
  paramsGE = data.table::data.table(NULL),
  minimize = TRUE,
  repls = 25,
  showInfo = TRUE,
  ncpus = NA,
  seed = 1
)
}
\value{
A plot containing one boxplot curve for each configurations benchmarked.
}
\description{
This functions benchmarks the
then plots them as boxplots wrt to their iterations.
}
\references{
Bernd Bischl, Jakob Richter, Jakob Bossek, Daniel Horn, Janek Thomas and Michel Lang; mlrMBO: A Modular Framework for Model-Based Optimization of Expensive Black-Box Functions, Preprint: \code{\link{https://arxiv.org/abs/1703.03373}} (2017).
}
