% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task}
\alias{task}
\title{Generate task}
\usage{
task(simulation = "regr.randomForest", data, target, psOpt, minimize = TRUE)
}
\arguments{
\item{simulation}{[\code{character}]\cr
The black box function e.g. model for the \code{mlrMBO}
Default is `regr.randomForest`.}

\item{data}{[\code{data.frame}]\cr
Data of problem. \cr}

\item{target}{[\code{character}]\cr
The variable, which one wants to minimize (maximize) \code{mlrMBO}.}

\item{psOpt}{[\code{ParamHelpers::ParamSet()}]\cr
Collection of parameters and their constraints for problem optimization}

\item{minimize}{[\code{logical(1)}]\cr
Should the target be minimized? \cr
Default is `TRUE`.}

\item{task}{[\code{EBO:: task()}]\cr
Task defines the problem setting.}
}
\value{
List with relevant informations of a specific problem.
}
\description{
This functions enables users to create a task, which mainly describes the optimization problem.
Task is required for many EBO functions, such as EBO::optimizertuneRace(), EBO::generateConfigdata()
or EBO::generateProblemList().
}
