% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneMboMbo.R
\name{tuneMboMbo}
\alias{tuneMboMbo}
\title{Benchmark and plot \code{mlrMBO::mbo()} optimization runs to investigate hyperparameters}
\usage{
tuneMboMbo(
  instance,
  psOpt,
  funcEvals,
  psTune,
  itersMboTune = 10,
  minimize = FALSE,
  repls = 15,
  ncpus = NA,
  seed = 1,
  designOpt = NULL,
  maxTime = NULL
)
}
\description{
This functions benchmarks the \code{mlrMBO::mbo()} function on different configurations and
then plots them wrt the hyperparameters.
}
